Atv.3.4/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
gensim==4.3.3
h11==0.16.0
idna==3.10
joblib==1.5.1
nltk==3.9.1
numpy==1.26.4
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
regex==2024.11.6
scikit-learn==1.7.0
scipy==1.13.1
smart-open==7.1.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
wrapt==1.17.2
